refactor mml_player
