+++ lib_player/src/lib.rs
@@ -4,6 +4,7 @@ mod note_event;
 mod mml_event;
 mod utils;
 mod mml_player;
+mod track_player;
 
 pub use synth::Synth;
 pub use synth::SynthOutputConnection;
@@ -11,6 +12,8 @@ pub use synth::SynthOutputConnection;
 pub use mml_player::MmlPlayer;
 pub use mml_player::MmlPlayerOptions;
 pub use mml_player::NoteOnCallbackData;
+pub use mml_player::PlaybackStatus;
 
 pub use note_event::NoteEvent;
 pub use parser::Parser;
+pub use track_player::TrackPlayer;
